Hey Yannis, check description for comments
Under task 4 I inserted how I think we should have some of the unknown variables we have talked about today. The number 1000 is just a random number for the size of each variable. I think the thing of you to check tomorrow is to set up the neural network in order to accept values of u, σu and α and get values for the moments. Maybe check after you train the NN model if there is a way to give like Xtest which will include these three uncertain variables and give us the moment. We will catch up tomorrow.